% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_server.R
\name{createAcquacottaServer}
\alias{createAcquacottaServer}
\title{The server function to pass to \link[shiny]{shinyApp}}
\usage{
createAcquacottaServer(config = createConfig())
}
\arguments{
\item{config}{Default parameters. By default generated by \link[acquacotta]{createConfig}.}
}
\description{
The server function to pass to \link[shiny]{shinyApp}
}
