% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_server.R
\name{runAcquacotta}
\alias{runAcquacotta}
\title{Run an Acquacotta instance}
\usage{
runAcquacotta(config = createConfig())
}
\arguments{
\item{config}{Default parameters. By default generated by \link[acquacotta]{createConfig}.}
}
\description{
Run an Acquacotta instance
}
